requirements.txt
Click==7.0
gitdb2==2.0.5
GitPython==2.1.11
Jinja2==2.10
livereload==2.5.2
Markdown==3.0.1
MarkupSafe==1.0
mkdocs==1.0.4
mkdocs-material==3.0.6
Pygments==2.2.0
pymdown-extensions==6.0
PyYAML==3.13
six==1.11.0
smmap2==2.0.5
tornado==5.1.1
